+++ src/test/kotlin/me/bossm0n5t3r/leetcode/bestTimeToBuyAndSellStock2/BestTimeToBuyAndSellStock2Test.kt
@@ -1,10 +1,10 @@
-package bestTimeToBuyAndSellStock2
+package me.bossm0n5t3r.leetcode.bestTimeToBuyAndSellStock2
 
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
 
 class BestTimeToBuyAndSellStock2Test {
-    private val bestTimeToBuyAndSellStock2 = BestTimeToBuyAndSellStock2()
+    private val bestTimeToBuyAndSellStock2 = BestTimeToBuyAndSellStock2.Solution()
 
     data class BestTimeToBuyAndSellStock2TestData(
         val prices: IntArray,
